requirements.txt
Flask==3.0.2
pypdf==4.0.1
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0